svc: Use proper random entropy generation algorithm

+++ src/core/settings.h
@@ -115,7 +115,7 @@ struct Values {
     // System
     bool use_docked_mode;
     bool enable_nfc;
-    std::optional<u64> rng_seed;
+    std::optional<u32> rng_seed;
     s32 current_user;
     s32 language_index;
